requirements.txt
# Pipeline
requests==2.31.0
duckdb==0.10.0
python-dotenv==1.0.0

# Orchestration
prefect==2.19.0

# Transform
dbt-core==1.8.0
dbt-duckdb==1.8.0

# Dev
pytest==8.0.0
